first pass at dynamodb persister
persist to a user provided table. There are some assumptions on the
table: it must exist, contain 'task_id' as a primary key and 'timestamp'
as a rangekey, such that it is the pairing of task_id and timestamp that
acts as a composite primary key for a given row in the table.

There is also some issues with serialization: we produce a dict on
deserialization, rather than the pyrsistent data-structure that is
serialized.