Make the LIKE constraint a "full match"
(COMPINFRA-333) We have a "stable" pool and a "stable_batch" pool, and
jobs that are supposed to run on the "stable" pool are getting scheduled
on the "stable_batch" pool because the constraint is too open. This
change should make it so that stuff only gets scheduled on the pool they
specify.