Add `stopping` instance variable to handle thread restarts
maksym pointed out that it's possible for one of our threads to
encounter an error and leave task_processing in a non-working state
since queues will either not be populated or never be consumed from.

to resolve this, i've added a `stopping` variable that gets set when
stop() is called and had both of our threads check those. in addition,
the Pod event processing loop will also attempt to continue processing
while its queue still has data in it to ensure that we cleanup as much
as possible before shutting down.