Add support for overriding resource requests (#207)
Resource requests are currently not explicitly set - this results in k8s setting
the memory/cpu/disk requests to the same value as the limit.

For the most part, this is desired behaviour at Yelp as the main consumer of this
library is Tron and we don't expect our batches to need to "burst" any of these
resources. Additionally, we generally want to signal to k8s that our batches shouldn't
be killed if there are other (more interrupt-tolerant) workloads on an overloaded
node by defaulting the QoS class for any task_proc-created Pods to `Guaranteed`
precisely by setting `requests = limits`.

However, one of our other large task_proc workloads (Jolt) would like to have 
control over the requests as their workload runs in an isolated pool and cannot (at
least at the moment) be easily right-sized for optimal bin-packing/resource usage.

But, given that Jolt runs in its own pool, we can be a little cheeky and set lower
requests to pack Pods a little tighter since Jolt *is* quite interrupt-tolerant and
as such is resilient to its Pods getting killed due to node resource pressure (which
should be pretty rare in any case given the current usage patterns).